reinit only reserves new memory if size changed

+++ include/tdp/data/managed_image.h
@@ -30,6 +30,8 @@ class ManagedImage : public Image<T> {
   }
 
   void Reinitialise(size_t w, size_t h) {
+    if (this->w_ == w && this->h_ == h)
+      return;
     if (this->ptr_)  {
       Alloc::destroy(this->ptr_);
     }
